reworked file management for profile setup
