+++ profile.html
@@ -433,6 +433,6 @@
       >
     </footer>
     <script src="/node_modules/bootstrap/dist/js/bootstrap.bundle.min.js"></script>
-    <script src="/src/js/profile/profile.mjs" type="module"></script>
+    <script src="/src/js/index.mjs" type="module"></script>
   </body>
 </html>
