+++ src/js/index.mjs
@@ -4,6 +4,7 @@ import { registerUser } from "./auth/registration.mjs";
 import { loginUser } from "./auth/login.mjs";
 import { displaySpecificPost } from "./posts/getSpecificPost.mjs";
 import { setup } from "./posts/setup.mjs";
+import { setupProfile } from "./profile/setup.mjs";
 
 const path = window.location.pathname;
 console.log(path);
@@ -16,4 +17,6 @@ if (path === "/index.html") {
   setup();
 } else if (path === "/post-specific.html") {
   displaySpecificPost();
+} else if (path === "/profile.html") {
+  setupProfile();
 }
